api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
redis==5.0.8
celery==5.4.0
pandas==2.2.2

